modularizing DeleteConfirm modal
