+++ client/src/components/RegularPalaceView.jsx
@@ -8,10 +8,10 @@ import {
   FaCheck,
   FaTimes,
 } from "react-icons/fa";
-import { AiOutlineMinusCircle } from "react-icons/ai";
 import AlertMessage from "./AlertMessage";
 import RoomView from "./RoomView";
 import PalaceCoverImage from "./PalaceCoverImage";
+import DeleteConfirm from "./DeleteConfirm";
 
 function RegularPalaceView({
   setShowAppAlert,
@@ -65,11 +65,6 @@ function RegularPalaceView({
     window.add_room_view.showModal();
   };
 
-  //on palace delete cancel
-  const handleCancelDelete = () => {
-    window.palace_delete_confirm.close();
-  };
-
   //on palace delete confirm
   const handleConfirmDelete = async () => {
     window.palace_delete_confirm.close();
@@ -110,35 +105,7 @@ function RegularPalaceView({
   };
   return (
     <>
-      {/*delete confirm modal*/}
-      <dialog
-        id="palace_delete_confirm"
-        className="modal m-auto w-1/4 min-w-fit  text-gray-600"
-      >
-        <form method="dialog" className="modal-box bg-gray-100 text-center">
-          <button className="btn btn-circle btn-ghost btn-sm absolute right-2 top-2">
-            ✕
-          </button>
-          <h3 className="mb-4 text-lg font-bold">
-            Are you sure you want do delete this palace?
-          </h3>
-          <div className="confirm-selection">
-            <button
-              className="btn mr-2 bg-gray-200 hover:bg-gray-300"
-              onClick={handleCancelDelete}
-            >
-              Cancel
-            </button>
-            <button
-              className="btn bg-red-500 text-white hover:bg-red-600"
-              onClick={handleConfirmDelete}
-            >
-              Delete
-            </button>
-          </div>
-        </form>
-      </dialog>
-
+      <DeleteConfirm handleConfirmDelete={handleConfirmDelete}></DeleteConfirm>
       <dialog id="reg_view" className="modal">
         <form method="dialog" className="modal-box">
           <AlertMessage
@@ -223,22 +190,22 @@ function RegularPalaceView({
                         </div>
                       ) : (
                         <div className="flex opacity-60">
-                          <span
-                            className="rounded px-2 py-1 text-lg text-white duration-200 hover:text-2xl hover:ease-in-out"
+                          <div
+                            className="rounded px-2 py-1 text-lg text-white duration-200 hover:scale-150 hover:ease-in-out"
                             onClick={() => {
                               handleRoomClick(room._id);
                             }}
                           >
                             <FaRegEye />
-                          </span>
-                          <span
-                            className="rounded px-2 py-1 text-lg text-white duration-200 hover:text-2xl hover:ease-in-out"
+                          </div>
+                          <div
+                            className="rounded px-2 py-1 text-lg text-white duration-200 hover:scale-150 hover:ease-in-out"
                             onClick={() => {
                               handleRoomEdit(room._id);
                             }}
                           >
                             <FaEdit />
-                          </span>
+                          </div>
                         </div>
                       )}
                     </div>
